+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/util/JudgeDecisionHelperTest.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.divorce.orchestration.util;
 
+import org.hamcrest.core.Is;
 import org.junit.Test;
 
 import java.util.HashMap;
@@ -8,6 +9,7 @@ import java.util.Map;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.JUDGE_COSTS_CLAIM_GRANTED;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.CcdFields.JUDGE_COSTS_DECISION;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NO_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.util.JudgeDecisionHelper.isJudgeCostClaimAdjourned;
@@ -51,6 +53,27 @@ public class JudgeDecisionHelperTest {
     public void isJudgeCostClaimEmptyReturnsFalse() {
         assertThat(isJudgeCostClaimEmpty(createCaseData(JUDGE_COSTS_CLAIM_GRANTED, "SomeValueThatIsNotNull")), is(false));
     }
+
+    @Test
+    public void givenJudgeCostDecisionNo_whenHasJudgeMadeCostsDecision_thenReturnFalse() {
+        Map<String, Object> caseData = new HashMap<>();
+        caseData.put(JUDGE_COSTS_DECISION, "No");
+        assertThat(JudgeDecisionHelper.hasJudgeGrantedCostsDecision(caseData), Is.is(false));
+    }
+
+    @Test
+    public void givenJudgeCostDecisionEmpty_whenHasJudgeMadeCostsDecision_thenReturnFalse() {
+        Map<String, Object> caseData = new HashMap<>();
+        assertThat(JudgeDecisionHelper.hasJudgeGrantedCostsDecision(caseData), Is.is(false));
+    }
+
+    @Test
+    public void givenJudgeCostDecisionYes_whenHasJudgeMadeCostsDecision_thenReturnTrue() {
+        Map<String, Object> caseData = new HashMap<>();
+        caseData.put(JUDGE_COSTS_DECISION, "Yes");
+        assertThat(JudgeDecisionHelper.hasJudgeGrantedCostsDecision(caseData), Is.is(true));
+    }
+
 
     private static Map<String, Object> createCaseData(String field, Object value) {
         Map<String, Object> caseData = new HashMap<>();
